chore: major refactor
