--- packages/backend/src/apps/custom-api/common/custom-http-client.ts
@@ -1,128 +1,0 @@
-import { IHttpClientParams } from '@plumber/types'
-
-import axios, { InternalAxiosRequestConfig } from 'axios'
-import { URL } from 'url'
-
-import HttpError from '@/errors/http'
-import RetriableError from '@/errors/retriable-error'
-import { processUrlPathParams } from '@/helpers/http-client/process-url-path-params'
-
-import { streamResponse } from './stream-response'
-
-export { AxiosInstance as IHttpClient } from 'axios'
-
-const removeBaseUrlForAbsoluteUrls = (
-  requestConfig: InternalAxiosRequestConfig,
-): InternalAxiosRequestConfig => {
-  try {
-    const url = new URL(requestConfig.url)
-    requestConfig.baseURL = url.origin
-    requestConfig.url = url.pathname + url.search
-
-    return requestConfig
-  } catch {
-    return requestConfig
-  }
-}
-
-export default function createCustomApiHttpClient({
-  $,
-  baseURL,
-  beforeRequest,
-  requestErrorHandler,
-}: IHttpClientParams) {
-  const instance = axios.create({
-    baseURL,
-    allowAbsoluteUrls: false,
-    // Security: Enable streaming to handle large responses safely
-    responseType: 'stream',
-    // Security: Set reasonable timeout to prevent hanging requests
-    timeout: 60000, // 60 seconds
-  })
-
-  // Edge case: unlike response interceptors, axios request interceptors are
-  // currently processed in a LIFO order. To prevent confusion, we'll only use 1
-  // request interceptor.
-  instance.interceptors.request.use(
-    async (
-      requestConfig: InternalAxiosRequestConfig,
-    ): Promise<InternalAxiosRequestConfig> => {
-      let newRequestConfig = processUrlPathParams(requestConfig)
-      newRequestConfig = removeBaseUrlForAbsoluteUrls(newRequestConfig)
-
-      // Intentionally serial as each callback works on result of previous one.
-      for (const callback of beforeRequest) {
-        newRequestConfig = await callback($, newRequestConfig)
-      }
-
-      return newRequestConfig
-    },
-  )
-
-  instance.interceptors.response.use(
-    (response) => streamResponse(response),
-    async (error) => {
-      // EDGE CASE: We allow actions / triggers to throw RetriableError, and
-      // some of them may choose to throw from beforeRequest. If this happens,
-      // Axios will still process the error through this response interceptor.
-      //
-      // Since RetriableError is supposed to be passed directly through to
-      // BullMQ, we special case it here instead of converting it to HttpError.
-      //
-      // Note that we still want to convert _other_ error types (even if
-      // they're not axios errors!) to HttpError; this helps prevent accidental
-      // sensitive data leakage via error objects in our logs.
-      if (error instanceof RetriableError) {
-        throw error
-      }
-
-      // This handles system errors like ECONNREFUSED, which don't have a
-      // response body.
-      if (!error.response) {
-        throw new HttpError(error)
-      }
-
-      const { config } = error
-      const { status } = error.response
-
-      if (
-        (status === 401 || status === 403) &&
-        $.app.auth.refreshToken &&
-        !$.app.auth.isRefreshTokenRequested
-      ) {
-        $.app.auth.isRefreshTokenRequested = true
-        await $.app.auth.refreshToken($)
-
-        // retry the previous request before the expired token error
-        const newResponse = await instance.request(config)
-        $.app.auth.isRefreshTokenRequested = false
-
-        return newResponse
-      }
-
-      throw new HttpError(error)
-    },
-  )
-  // We use a separate interceptor for requestErrorHandler to allow the above
-  // HttpError inteceptor to throw early.
-  instance.interceptors.response.use(
-    (response) => response,
-    async (error) => {
-      if (!requestErrorHandler) {
-        throw error
-      }
-
-      // Passthrough other errors... although other errors should really only be
-      // RetriableError.
-      if (!(error instanceof HttpError)) {
-        throw error
-      }
-
-      await requestErrorHandler($, error)
-
-      throw error
-    },
-  )
-
-  return instance
-}
