+++ packages/backend/src/helpers/http-client/index.ts
@@ -5,9 +5,10 @@ import { IHttpClientParams } from '@plumber/types'
 
 import { URL } from 'url'
 
-import createCustomApiHttpClient from '@/apps/custom-api/common/custom-http-client'
 import HttpError from '@/errors/http'
 import RetriableError from '@/errors/retriable-error'
+import { streamResponse } from '@/helpers/http-client/stream-response'
+import logger from '@/helpers/logger'
 
 import { processUrlPathParams } from './process-url-path-params'
 
@@ -31,22 +32,18 @@ export default function createHttpClient({
   beforeRequest,
   requestErrorHandler,
 }: IHttpClientParams) {
-  // Custom API special case: create a different http client for custom api
-  // where we stream the response to protect against gzip bombs
-  // other http clients should not be affected by this as they
-  // are controlled URLs from actions such as FormSG or webhook triggers
-  if ($?.app?.key === 'custom-api') {
-    return createCustomApiHttpClient({
-      $,
-      baseURL,
-      beforeRequest,
-      requestErrorHandler,
-    })
-  }
-
   const instance = axios.create({
     baseURL,
     allowAbsoluteUrls: false,
+
+    // Custom API special case: create a different http client for custom api
+    // where we stream the response to protect against gzip bombs
+    // other http clients should not be affected by this as they
+    // are controlled URLs from actions such as FormSG or webhook triggers
+    ...($?.app?.key === 'custom-api' && {
+      responseType: 'stream',
+      timeout: 60000, // 60 seconds
+    }),
   })
 
   // Edge case: unlike response interceptors, axios request interceptors are
@@ -69,7 +66,8 @@ export default function createHttpClient({
   )
 
   instance.interceptors.response.use(
-    (response) => response,
+    (response) =>
+      $?.app?.key === 'custom-api' ? streamResponse(response) : response,
     async (error) => {
       // EDGE CASE: We allow actions / triggers to throw RetriableError, and
       // some of them may choose to throw from beforeRequest. If this happens,
@@ -89,6 +87,18 @@ export default function createHttpClient({
       // response body.
       if (!error.response) {
         throw new HttpError(error)
+      }
+
+      // Process error response stream to parse JSON error body
+      if (
+        error.response.data &&
+        typeof error.response.data.pipe === 'function'
+      ) {
+        try {
+          await streamResponse(error.response)
+        } catch (streamError) {
+          logger.warn('Error processing error response stream:')
+        }
       }
 
       const { config } = error
