+++ packages/backend/src/helpers/http-client/stream-response.ts
